requirements.txt
Package                 Version
opencv-python	        4.5.1.48
pip	                    21.0.1
numpy	            	1.20.1
pdf2image		        1.14.0
pytesseract	 	        0.3.7
azure-storage-blob	    12.8.0
pyodbc	                4.0.30